+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## 1.12.0 [2022-01-19]
+
+### Features
+1. Add delete method for BaseModel
+
 ## 1.11.0 [2022-01-09]
 
 ### Features
-1. Fix issue #8: Migrations on standalone Clickhouse database 
+1. Fix issue #8: Migrations on standalone Clickhouse database
 
 ## 1.10.0 [2021-09-20]